// apps/creator-hub/modules/creations/placeThumbnails/types.ts
export enum MediaType {
  Image = 'Image',
  Video = 'Video',
  AutoGeneratedImage = 'AutoGeneratedImage',
}

export type Thumbnail = {
  /// thumbnailId as a field will be removed with enablePlaceThumbnailAssetsUploadApi flag removal
  thumbnailId: number;
  imageId: number;
  type: MediaType;
  videoHash?: string;
  altText?: string;
  videoTitle?: string;
  assetId: number;
};